[op][n_upsable][ozone][cursor] Set initial value of custom cursor to "default"
:Release Notes:
Set initial value of custom cursor to "default"

:Detailed Notes:
Before this patch, the initial value for the cached cursor type in
OzoneWaylandWindow was kBlank (blank cursor or disabled in LSM). We need
to avoid an scenario that is happening often in chromium72, where app window
is focused for the first time before before any mouse event or initialization
happens, and ends up disabling cursor. Once cursor is disabled, no event
will make it show again, unless it is explicitely requested through
injection.

The solution is setting the initial value to kNotUse (equivalent to enabling
default cursor).

:Testing Performed:
1. Open CMS player
2. When loading, put the cursor over the position of the filter/sort
   button
3. Do not move the cursor and wait till it goes away.
4. Now cursor always comes back when you hover it again.

:QA Notes:
Test-cases: QS-7717, GT-23503, GT-25241

:Issues Addressed:
[PLAT-84296] CMS Player single image 12 hour playback is getting failed
[PLAT-98671] Port patch: [open][neva_upstreamable][ozone][cursor] Set initial value of custom cursor to "default"

Change-Id: Id1b10ddbbb86765c3a2cc0534b951dcb7fc7b4e6
Originally-Reviewed-on: http://gpro.lge.com/252620
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/266906
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>