[op][n_upsable][dynamic_sized_window] Implement dynamic sized window
:Release Notes:
Implement dynamic sized window
Upscale video to fullscreen providin additional contents scale to media
stack. Keep size of size-overriding apps when relaunched.

:Detailed Notes:
If the application's actual resolution is different from platform resolution,
then WAM & webOS Webview adjusts app's content size to fit the wayland window.
It, however, is redundant to create a larger window than app's resolution,
in terms of WAM's memory usage.

Moreover, additional cost may be needed to scale content up.
This patch makes wayland window have app's actual resolution.
In webOS, system compositor scales window to match platform
screen size when window is displayed on screen.

Playing video with webos media stack it is needed to provide the target position for
the video in UI resolution (and indicate if it is fullscreen or not). For
that, the algorithm uses additional contents scale, to adapt the media
resolution to the UI resolution.
This is needed for properly showing fullscreen videos on videos that
override the application resolution at fullscreen (regular way to see apps
in webOS). The window will get the overriden resolution, but it is
upscaled to show fullscreen. Videos need to adapt to the expected
resolution.

Sometimes on relaunching, the wayland window (and
app size) was resized to display screen size, after fullscreen.
There were two problems. One was that on setting the window fullscreen,
the bounds were set to the size of the display. This has been fixed
implementing the ToggleFullscreenWithSize method.

The other problem was that, once application got fullscreen status,
window manager was sending a configure event with the display size.
Solution is - ignoring this. For that this time
we use the WidgetDelegate::CanResize(), that is set
to false in app runtime already. Additionally this patch adds the
handling of the resize behavior that is populated using CanResize
value.

:Testing Performed:
Test1:
1. Install test.resolution.1280-720_1.0.0_all.ipk from wiki/x/hxG0C
2. launch the *test.resolution.1280-720* app
   luna-send -n 1 -f luna://com.webos.applicationManager/launch '{"id": "test.resolution.1280-720"}'
3. Ensure app is displayed with full screen and 720P logo is fully
visible

Test2: GT-20753
1. Launch following application (their resolution in appinfo.json is
1280x720):
 - Dailymotion
 - NFB Films
 - Starz
 - Arirang
 - iHeartRadio
2. Ensure that application is fullscreen and scaled correctly

Test3: GT-15269
1. Launch NFB Films or any app which is not screen size (app with resolution 1280x720)
2. Play embedded or fullscreen video
3. Confirm video scaling & position is correct

Test4: GT-27618
1. Launch NFB Films (view is shown properly, taking all the screen).
2. Goto launcher and launch NFB films again.
3. Confirm the application keeps showing fullscreen.

Test5: GT-27617
1. Launch NFB Films (view is shown properly, taking all the screen).
2. Play fullscreen video
3. Then switch to another application (photo-video)
4. Switch back to NFB films.
5. Confirm the video scaling & position is correct

Test6: GT-23959
1. Remove the ethernet cable - No network on device
2. Launch Arirang
3. Confirm Network error page is shown in proper resolution & scaling

:QA Notes:
Test-cases: GT-20753, GT-15269, GT-27617, GT-23959, GT-27618

:Issues Addressed:
[PLAT-98659] Port patch: [open][neva_upstreamable][dynamic_sized_window] Implement dynamic sized window
[PLAT-98783] [chromium79] Full screen display failed when running 1280*720 app.

Change-Id: I4aee8d35a06f4a0651dcdd88a3d9273d370b82d2
Originally-Reviewed-on: http://gpro.lge.com/245585/
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/267483
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Reviewed-by: OSE Helper Bot <ose@lge.com>