[oe][ds][media][ls2] Match URIs array with URITypes
:Release Notes:
Match URIs array with URITypes

:Detailed Notes:
luna_service_uris array is indexed based on URIType, but there was
a mismatch between the sizes of each.
As some services are not available in OSE, use empty strings so that
there's no buffer overrun when accessing it based on any URIType.

:Testing Performed:
[QO-1043] Basic testing of Blink BAT for Open on the RPI target

:QA Notes:
Test-cases: [QO-1043]

:Issues Addressed:
[PLAT-98594] Port buffered range for url media
[NEVA-4464] Port Remove unsupported mimes and codecs for OSE

Change-Id: Ica1ef602125ce2f097904aa1bd7804be95b3d16d
Originally-Reviewed-on: http://gpro.lge.com/245671
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/267233
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>