[op][ds][cbe] Do not flag cbe or app_shell as testonly
:Release Notes:
Do not flag cbe or app_shell as test only.

:Detailed Notes:
When using enactbrowser, app_shell is our production browser, so it should
not have any test library or component. So we need to modify the original
app_shell testonly flags to be unset. Code needs to be modified so a lighter
version of content_shell_lib (named conntent_shell_nontest_lib) can be
used from app_shell.

:Testing Performed:
Build. Use strings on libcbe.so to check ShellBrowserMain is not present.

:QA Notes:
Test-cases: GT-26820

This is not reducing size back to 70M as we had in chromium68. Mojoification
has made Chromium bigger in general. Though, with this change we drop around
150K that is the implementation of the content shell test components.

In chromium79 libcbe.so size is around 80MB after applying this.

:Issues Addressed:
[PLAT-81235] [chromium72] libcbe.so disk size is bigger than 70MB
[PLAT-98720] Port patch: [open][downstream][cbe] Do not flag cbe or app_shell as testonly

Change-Id: I43902c79cb48a2272c886cdf5c792e2acb718902
Originally-Reviewed-on: http://gpro.lge.com/247889
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/267820
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>