[op][ds][layout][webos] Do not implicitely set min-height: auto for flexbox
:Release Notes:
Keep legacy behavior of column flexbox flows in a flex layout

:Detailed Notes:
After upgrading to chromium79, firstuse pickers and license panel
have started to fail. The reason is a change in the default behavior
for flexbox children inside a flex layout. Instead of implicitely
setting min-height: 0px now it sets min-height: auto as is expected
in standard.

The way to avoid the need for this workaround is setting min-height
to 0 when needed. And this has been already done in Enact 3.x. But
legacy Enact versions (as 2.0 version used in firstuse) still don't
have the fix.

This fix to offer the old behavior should be dropped as soon as
firstuse moves to use a version of enact with the fix.

:Testing Performed:
1. Launch first use
2. On location page, open language picker. It should only take a small
part of the screen instead of showing the countries mixed with the
other pickers. Scrollable buttons should be visible.
3. On EULA page, license should show scroll buttons, and they should be
scrollable.

:QA Notes:
Test-cases: GT-28156

:Issues Addressed:
[PLAT-98624] [chromium79] FirstUse app: Some elements are not visible.

Change-Id: I313986622aee7b2d41b2fe50151f06a221101f63
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/268595
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>