[op][n_upsable][app_runtime] Dispatch visibility change event on page paused state
:Release Notes:
Dispatch visibility change event on page paused state

:Detailed Notes:
There was a problem that event handlers on 'onvisibilitychange' or
'onwebkitvisibilitychange' are not working when a webapp is resumed
by app switching. It happened when the visibie signal is dispatched
only before the page is resumed, so makes it ignored in that case.

The problem had been handled in upstream codes, but now it has been
covered in Neva app runtime due to not upstreamable. I couldn't find
any relevant problem on x86 vanilla Chrome, so it couldn't be
acceptable in upstream. In addition, the root cause has been derived
from WAM in which inappropriate visibility state signal order were
delivered. Thus we handle the situation in Neva app runtime regardless
of the visibility state signal order sent from app runtime clients.

:Testing Performed:
Ran GT-21986 and GT-14761, chromium79 build
Test 1:
1. Launch https://codepen.io/anawhj/pen/JjoxJQz
2. Switch to another app
3. Switch back to the codepen app launched first
4. check the visible and hidden messages are shown on the page

:QA Notes:
Test-cases: GT-21986, GT-14761

:Issues Addressed:
[PLAT-98592] Port Dispatch visibility change event on page paused state
[PLAT-81252] [chromium72] LG store launches but resuming of AD is not played.
[PLAT-83382] Youtube video is not played after switching

Change-Id: I9fa328278037785428d36558f96c2f9d1085ef04
Original-Reviewed-on: http://gpro.lge.com/250543
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/268776
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>