[op][ds][splash] Send FMP on blink swap, and not on frame presented.
:Release Notes:
Send FMP on blink swap and not on frame presented.

:Detailed Notes:
Upstream chromium has introduced a change where the performance timings
related with frames exposed to user should account the time the frame
is actually presented to user. But splash screen depends on FMP to
start swapping frames, so this is actually not happening if swap
promises are pending when FMP is required.

To make splash work again as expected regarding getting FMP, we
just make the swap events be sent as in the past, as they come
from FMP detector, and not when presentation happens.

The problem is introduced by this change:
    commit 992979a662b28e7b9a62419f87b1839557611efc
    Author: Sadrul Habib Chowdhury <sadrul@chromium.org>
    Date:   Wed Jun 13 19:16:01 2018 +0000
    Reland "paint timing: Use presentation-time instead of swap-time."

:Testing Performed:
Test 1:
1. Launch userguide from luna command (or from recent apps after it
has been closed
2. Splash should be dismissed in a few seconds.

Test 2:
1. Launch 1280x720 test app
2. Check it is rendered properly.

:QA Notes:
Test-cases: GT-28157, QS-8559

:Issues Addressed:
[PLAT-98783] [chromium79] Full screen display failed when running 1280*720 app.
[PLAT-100146] [chromium79] App relaunch doesnt work for tvuserguide app when launched from recents

Change-Id: I4878bda67d2a762ded7d874069cba1c34e6a932f
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/268684
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>