[op][n_upsable][injections] Do not keep a singleton of WebOSServiceBridgeProperties
:Release Notes:
Do not keep a singleton of WebOSServiceBridgeProperties

:Detailed Notes:
When a web application creating different frames, some of them that
could go away, is created, we reset the WebOSServiceBridgeProperties
singleton for each extension we install. We are installing one injection
per frame. So it could happen that the last creater properties object
would make reference to a frame that is not existing anymore.

In that case there would not be a counterpart in the browser process
so the actual call to check the identifier would return an empty value.

To avoid this, we just create the properties object, and fetch the
application ID identifier when the actual injection object is created.

:Testing Performed:
Run automated Youtube test

:QA Notes:
Test-cases: P2 Cat 11 youtube test

:Issues Addressed:
[PLAT-100159] [chromium79] CAT youtube test failure - Test Automation

Change-Id: I6e3f91d64349254b614e25192190727271854cd0
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/268956
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>