[op][n_upsable][file_access][whitelisting] Use whitelist logic also for directory navigation
:Release Notes:
Directories should be also checked using whitelisting logic.

:Detailed Notes:
Neva adds a new method to validate file access that is the one used
to delegate to AppRuntimeFileAccessDelegate. We were using it
for file access, but not for directory access.

:Testing Performed:
1. Launch enyo2sampler or user guide
2. Sideload on web inspector window.location="/etc"
It should show an error instead of listing the contents.

:QA Notes:
Test-cases: GT-18780, GT-18781

:Issues Addressed:
[PLAT-100581] [chromium79] File WhiteListing feature broken on chr79

Change-Id: I05f086e2cf60740bdd94e330a85ba130ef81fc6b
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/269487
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>