[op][n_ups][upsable][memory] Clear pending BackgroundTracingAgentClientImpl constructor
:Release Notes:
Clear pending BackgroundTracingAgentClientImpl constructor.

:Detailed Notes:
When renderer process is created, BackgrundTracingManagerImpl creates
closure for contructing of BackgroundTracingAgentClientImpl and creates
its instance lazily only when any agent observers are registered.
But it never releases if non of agent observer exists so it remained
as leaked memory.

To release pending constructor, BackgroundTracingAgentClientImpl registers
a disconnect handler to BackgroundTracingAgentProvider then it releases
when it is disconnected.
BackgroundTracingAgentProvider would be disconnected when the process is gone.

:Testing Performed:
1. Modify content/browser/tracing/backgroud_tracing_manager_impl.cc:489
from
DVLOG(1) << __func__ << " # pending ctor="
to
LOG(INFO) << __func__ << " # pending ctor="

2. Redirect chromium log to /var/log/wam.log by adding below to
webapp-mgr.conf

exec > /var/log/wam.log 2>&1

Run '6000 bareapp launches & closes' and check log "# pending ctor="
When app is starting check below log and the number should not increase
steadily(the number is depend on the number of created processes)

AddPendingAgentConstructor # pending ctor=2

:QA Notes:
To enable debug log, source code needs to be re-compile and install
libcbe.so to /usr/lib folder.
Test-cases: N/A

:Issues Addressed:
[NEVA-4413] NEVA Memory usage based on chr.79 increased about 30MB
[PLAT-100262] [chromium79] Chromium79 uses more GPU memory compared to chromium72

Change-Id: I6ac5e0c3b553963b911eb29745df92b57416e598
Originally-Reviewed-on: http://repo.lge.com:8080/2940
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/270594
Reviewed-by: Build_Verification
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>