[op][n_upsable][file_access][whitelisting] Rework whitelisting to reject when not matched
:Release Notes:
Reject whitelist requests when not matched.

:Detailed Notes:
Reworked the code associates a file access request to a specific WebView state
(to access whitelisting rules for each web app), as it was not properly
resolving in the subresources requests, nor in child frames.

Also, made the code reject requests for file access if it would not be able
to resolve to a web view, to increase security.

Moved all the whitelisting logic to UI thread, so now WebOSWebViewRendererState
is not needed anymore. Now we resolve the requests directly to the WebViewBase
instance.

Finally, added LOG(WARNING) for the file access errors.

:Testing Performed:
Try loading /etc/ and /etc/passwd with redirection and xmlHttpRequest in:
- untrusted app
- trusted app
- iframe of untrusted app (enyo2sampler)

:QA Notes:
Test-cases: GT-18780, GT-18781, GT-18783, GT-18784

This breaks enyo2sampler/enact sampler media player as they try to
access zoneinfo, and that's not whitelisted.

:Issues Addressed:
[PLAT-100581] [chromium79] File WhiteListing feature broken on chr79

Change-Id: I51f49b31fb900c311441a13bad58920b4f5d8424
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/269488
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>