[op][upsable][wayland][touch] Use currentTime as timestamp for touch event
:Release Notes:
Use currentTime as timestamp for touch event

:Detailed Notes:
The web content compares the time stamp of the thumnails rendered
with the timestamp of click event fired to check the validation of
 click event.
To generate click event from gesture tap, we use the timestamp sent from
LSM. But the timestamp is not valid sometimes because of some reasons
like a latency of touch events between LSM and Chromium, LSM sends whole
lot different timestamp from the platform time because of the load of
the whole system. In detail, timestamp in wl_touch is the elapsed time
since LSM process started up.
So this commit sets the timestamp to currentTime for TouchUp and
TouchDown event instead of using it from wayland.

:Testing Performed:
Tested the web app related to this issue
with scrolling, selecting thumbnails
Verified click handler works properly.

:QA Notes:
Test-cases: QWA-2347

:Issues Addressed:
[PLAT-99804] Touch event not working for few seconds after
scrolling in the app

Change-Id: Ic8e2136cdb4bcf65524fb0fe54daca0fd4c5c13d
Originally-Reviewed-on: http://gpro.lge.com/269515
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/270017
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>