[op][n_upsable][media] Fix MediaSource.isTypeSupported error
:Release Notes:
Fix MediaSource.isTypeSupported to return correct data type

:Detailed Notes:
Fix MediaSource.isTypeSupported to return correct data type
In chromium79 MIMETypeRegistry::IsSupportedMediaSourceMIMEType is
returning bool but the internal function CheckTypeAndCodecs is
returning SupportsType enum.
So when we it return MayBeSupported for types which expects codec
but is not receiving one then from MediaSource::isTypeSupported it
is returning true. But it should be returning false.
For example for the types such as 'video/webm' or 'video/webm;' or
'video/webm;codecs' it returns true instead of returning false.
So changed the return type of these internal neva functions to
return SupportsType enum value instead of bool value.

:Testing Performed:
Local build is successfull.
1. Luanch some web app and connect to web inspector from desktop
   using debug mode
2. Lunch "https://w3c-test.org/media-source/mediasource-is-type-supported.html"
   using window.location
3. Test results for "video/webm" should be PASS

:QA Notes:
Test-cases: GT-28177, GT-28178, GT-28212

:Issues Addressed:
[PLAT-99399] syncup - [media][mse] Add webM HDR functions

Change-Id: Ibc8744b0055f243c505d442073b70a4fb69e8a1e
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/269936
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>