[op][n_upsable][net][static_analysis] Missing net::NetLog pointer initialization
:Release Notes:
Initialize net::NetLog pointer to nullptr.

:Detailed Notes:
This has been detected by static analysis. The pointer to net::NetLog
from neva_app_runtime::URLRequestContextFactory was missing an
initialization. In theory it should not be a problem as we don't use
it before the UI process initialization. But it is a risk if in the
future net::NetLog is used before.

:Testing Performed:
Build succesful.

:QA Notes:
Test-cases: not required (code quality fix).

:Issues Addressed:
[CID-8476797] Uninitialized pointer field

Change-Id: Ic2d5e295cf9e3915f4bbc7fcf7feb1c4286e93cb
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/271425
Reviewed-by: Build_Verification
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>