[oe][n_upsable][cookies] Store persistent cookies
:Release Notes:
Make WAM and Enact Browser store persistent cookies on the disk
by providing cookie paths and encryption configurations.

:Detailed Notes:
WAM and Enact Browser currently do not store persistent cookies
on the disk because no cookie path is provided
when the NetworkContext is created.
This patch modifies overriding methods of
ContentBrowserClient::CreateNetworkContext() to give a cookie path to
NetworkService::CreateNetworkContext().

This patch disables encryption of cookies for the time being.
Otherwise, the network process crashes with a null pointer dereference
trying to load a key storage before encrypting cookies.

:Testing Performed:
1. Launch the test app using the below command.
   luna-send -n 1 -f luna://com.webos.applicationManager/launch '
   {"id": "test.redirection", "params":
   {"target": "http://10.177.245.57/~gabwon.park/web/cookie.html"}}'
2. Fill out "TEST" in the input field "Enter name:"
   and click "Set Cookie."
3. Wait 30 seconds or more to make sure the cookie is stored
   on the disk.
4. Reboot the device.
5. Launch the test app using the below command.
   luna-send -n 1 -f luna://com.webos.applicationManager/launch '
   {"id": "test.redirection", "params":
   {"target": "http://10.177.245.57/~gabwon.park/web/cookie.html"}}'
6. Click "Get Cookie."

:QA Notes:
Test-cases: QWA-3651, QWA-3668

:Issues Addressed:
[PLAT-106004] Cookies file is not created in /var/lib/wam/Default

Change-Id: Icdfe8a601463684de3c16bd9be6e40cd2860b884
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/276075
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>