[op][n_upsable][app_shell][media] Pass media codec capability to renderer
:Release Notes:
Pass platform media codec capability to renderer preferences

:Detailed Notes:
Currently media codec capability defined by platform is not pass
to the renderer preferences for app_shell. After refactor of
MediaPlatformPrefes, initializing media backend is failed if we
don't pass media capability.
So with this changes we read the json content of the media
capability file into a string and set to the renderer preferences.

:Testing Performed:
1. Launch Enact Browser
2. Load http://www.youtube.com
3. Play any video.
4. Video plays properly

:QA Notes:
Test-cases: QWA-551, QWA-795, QWA-3439

:Issues Addressed:
[PLAT-108780] Video Does not play for particular websites in Enact
              Browser app
[PLAT-108886] YouTube video is not playing in Enact Browser.
[PLAT-108822] Enact Browser can't play YouTube site's contents.

Change-Id: I37a76c9deb7971a3ac53ae3015c571aa9c5ceb01
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/275720
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Alexander Trofimov <alexander.trofimov@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>