[oe][n_upsable][network_error] Fix to load YouTube test app with Tellurium
:Release Notes:
Fix to load YouTube test app with Tellurium

:Detailed Notes:
Currently YouTube test app is redirecting with changing window.location
but this occurs CORB error while loading cookie resource from
accounts.google.com.
The app itself can load normally with ignoring this error but
when tellurium is involved then there is a problem that tellurium is
using the appId as "settingsservice.client" since there was an error
previously.

But actually the error page was not shown and it was because
the error didn't occurred on the main frame.
So adding the condition for checking whether the error was on
the main frame or not, this will not set the appId to
"settingsservice.client" when the error page is not shown.

:Testing Performed:
1) Install telluriumnub & telluriumd
2) ls-monitor | grep tellurium
3) Launch YouTube test app
4) Wait until the YouTube page is fully loaded
5) Verify that com.webos.app.test.youtube is connected to telluriumd

:QA Notes:
Test-cases: QWA-3551, QWA-3815, QWA-3767, QWA-1055

:Issues Addressed:
[PLAT-107976] self.pytell_doc.get_elements_by_class_name(self.YOUTUBE_VIDEO_TITLE_CLASS) is returning None

Change-Id: I7253b930a5402c6f9668a39f3c672204d3d65b61
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/276172
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>