fixup! [op][n_upsable][media] refactor creating video window path
:Release Notes:
Check for a null pointer in
ForeignVideoWindowProvider::CreateNativeVideoWindow()
to make the static analyzer happy.

:Detailed Notes:
ForeignVideoWindowProvider::CreateNativeVideoWindow()
creates a ForeignVideoWindow in a map, sets a pointer to the window,
and uses the pointer to change some properties of the window.
This patch adds a null pointer check because the static analyzer thinks
the pointer to the window can be null.
Note that in reality an exception such as std::bad_alloc will be thrown
before the pointer becomes a null pointer.

This patch also fixes a typo in
PseudoVideoWindowProvider::CreateNativeVideoWindow().

:Testing Performed:
Buiid and check error messages.

:QA Notes:
Test-cases: GT-20750

:Issues Addressed:
[PLAT-109085] [coverity] Dereference null return value

Change-Id: Ic5e99c7d854d0a3e1bd20d4bf7d676827d420b1e
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/277155
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/277756
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>