fixup! [op][n_upsable][webrtc] Implement webrtc video encoder interface
:Release Notes:
Fix RTP timestamp issue for encoded frames in webrtc video encoder

:Detailed Notes:
Currently the incoming timestamp of raw video frames and the time
stamp of encoded frames are not considered to decide the actual
timestamp value to be used for an encoded frame to be streamed to
peer connection. As a result sometimes there is a big gap between
local preview and remote playback.
With this change we use the logic of RTCVideoEncoder upstream
implementation for deciding timestamp value for encoded timestamp,
and fix the issue.

:Testing Performed:
1. Tested Video conference with H264 & VP8 codec types using
   https://appr.tc/?vsc=H264 and https://appr.tc/?vsc=VP8
   in Auto board
2. Tested video conference with H264 codec type using
   https://appr.tc/?vsc=H264 in OSE board
3. Tested Video conference using https://meet.jit.si/
4. Tested streaming from canvas element using below url.
   https://webrtc.github.io/samples/src/content/capture/canvas-pc/

:QA Notes:
Test-cases: QWA-5260, QWA-2739, QWA-2743

:Issues Addressed:
[PLAT-123770] webRTC - "test.redirection.webrtc" app gets reloaded

Change-Id: I1628384606a6fbda5ad8d1c227ca48a88969307b
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/294803
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>