[op][n_upsable][wv_api] Remove ForceVideoTexture API
:Release Notes:
Remove ForceVideoTexture API as it is not used anywhere.

:Detailed Notes:
We are maintaining this ForceVideoTexture, apparently coming from
some code that is not actually leading anywhere right now. So we
drop it to simplify implementation for first frame detection code.

:Testing Performed:
Build completes.

:QA Notes:
Test-cases: not required (only build).
Depends on: http://gpro.lge.com/c/webosose/wam/+/289494

:Issues Addressed:
[PLAT-119527] Drop unused code for application life cycle that could make splash screen refactor harder

Change-Id: I380f7566c089860c74dae05a18fa7932ec88da76
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/289491
Tested-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>