[op][n_upsable][appruntime] New implementation for shifting of overlapped by VKB input field
:Release Notes:
Implemented a new way for shifting of overlapped by VKB input fields.

:Detailed Notes:
Implemented a new way for shifting of overlapped by VKB input fields.
Content will shift in delta height of overlapping and will not shift
if input field is going to be shifted out of screen.

:Testing Performed:
For Auto:
1. On AVN Monitor, navigate to MySettings->Profile->Add Profile
Touch on input field (AVN VKB will be invoked)
2. Touch "t" and "e"
3. Predictions bar on top of VKB should be invoked

For all platforms:
1. Launch com.palm.app.enyo2sampler
2. Goto Moonstone -> Input category
3. Click one of the element in the "Bottom-aligned Inputs".
Actual Result:
Viewport is shifted up so the input area is visible
4. Click outside of input field to hide VKB
Viewport is shifted down to the normal position

:QA Notes:
Test-cases: QWA-5617, QSDK-5557

:Issues Addressed:
[PLAT-120420] [Auto] Cursor Position is not updated after touching second character from AVN VKB

Change-Id: I996281aad66713fdf0fe59d2625d46f51cc71c4d
Reviewed-on: http://gpro.lge.com/c/webosose/chromium79/+/291412
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>